Homescreen Changes
This contains the HTML versions I have on my computer. I removed the background image because it conflicted with the input gui. I also added and created some basic pages for the Calendar, Profile, and Homepage. Don't put this on the main yet. Jason, if you could find a way to use java to get users login information, that would be great. - Justin